Camera, Bee and Berrys
Camera now changes along the Y axis with the worm as well as zooms in and out. Bee moves in a loop. Berrys bounce up and down, I could not figure out how to fix the player not moving glitch but it seems to work better when they are spaced out more.